迷路の表示まで実装した
