+++ src/init_app.rs
@@ -28,46 +28,30 @@ impl Plugin for Schedule
 
 ////////////////////////////////////////////////////////////////////////////////
 
-//text UIのメッセージセクションの型
-type MessageSect<'a> =
-(   &'a str, //表示文字列
-    &'a str, //フォントのAssets
-    f32,     //フォントのサイズ
-    Color,   //フォントの色
-);
-
-//ヘッダーのComponent
-#[derive( Component )] struct HeaderUiStages;
-#[derive( Component )] struct HeaderUiScore;
-#[derive( Component )] struct HeaderUiHiScore;
-
-//フッター(FPS表示)のComponent
-#[derive( Component )] struct FooterUiFps;
+//FPS表示のComponent
+#[derive( Component )] struct UiFps;
 
 //text UIの初期値
-// #[allow( dead_code )]
-pub const NA2  : &str = "##";
-pub const NA3  : &str = "###";
-pub const NA5  : &str = "#####";
-pub const NA2_5: &str = "##-#####";
-pub const NA3_2: &str = "###.##";
+const NA2  : &str = "##";
+const NA5  : &str = "#####";
+const NA2_5: &str = "##-#####";
+const NA3_2: &str = "###.##";
 
 //ヘッダーの設定
 counted_array!
-(   pub const TEXT_HEADER_LEFT: [ MessageSect; _ ] =
+(   const TEXT_HEADER_LEFT: [ MessageSect; _ ] =
     [   ( " STAGE ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.7, Color::GOLD  ),
         ( NA2      , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.7, Color::WHITE ),
     ]
 );
 counted_array!
-(   pub const TEXT_HEADER_CENTER: [ MessageSect; _ ] =
+(   const TEXT_HEADER_CENTER: [ MessageSect; _ ] =
     [   ( " SCORE ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.7, Color::GOLD   ),
         ( NA5      , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.7, Color::WHITE  ),
-        ( NA3      , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.5, Color::SILVER ),  //placeholder for debug
     ]
 );
 counted_array!
-(   pub const TEXT_HEADER_RIGHT: [ MessageSect; _ ] =
+(   const TEXT_HEADER_RIGHT: [ MessageSect; _ ] =
     [   ( " Hi-SCORE ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.7, Color::GOLD  ),
         ( NA5         , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.7, Color::WHITE ),
     ]
@@ -76,7 +60,7 @@ counted_array!
 //フッターの設定
 counted_array!
 (   const TEXT_FOOTER_LEFT: [ MessageSect; _ ] =
-    [   ( " FPS " , ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.60, Color::TEAL   ),
+    [   ( "  FPS ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.60, Color::TEAL   ),
         ( NA3_2   , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.40, Color::SILVER ),
         ( " demo ", ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID       * 0.45, Color::TEAL   ),
         ( NA2_5   , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.25, Color::SILVER ),
@@ -92,7 +76,7 @@ counted_array!
     [   ( "Powered by ", ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID * 0.6, Color::TEAL   ),
         ( "RUST"       , ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID * 0.6, Color::SILVER ),
         ( " & "        , ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID * 0.6, Color::TEAL   ),
-        ( "BEVY "      , ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID * 0.6, Color::SILVER ),
+        ( "BEVY  "     , ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID * 0.6, Color::SILVER ),
     ]
 );
 
@@ -159,90 +143,83 @@ fn spawn_text_ui
     asset_svr: Res<AssetServer>,
 )
 {   //レイアウト用の隠しフレームの準備
-    let per100 = Val::Percent( 100.0 );
     let width  = Val::Px( SCREEN_PIXELS_WIDTH  );
     let height = Val::Px( SCREEN_PIXELS_HEIGHT );
     let background_color = BackgroundColor ( Color::NONE );
 
-    let style = Style
-    {   width          : per100,
-        height         : per100,
-        position_type  : PositionType::Absolute,
-        flex_direction : FlexDirection::Column,
-//      justify_content: JustifyContent::Center,
-//      align_items    : AlignItems::Center,
+    let hidden_frame_header = NodeBundle
+    {   style: Style
+        {   width,
+            height,
+            position_type  : PositionType::Absolute,
+            flex_direction : FlexDirection::Column,
+            justify_content: JustifyContent::FlexStart, //画面の上端
+            ..default()
+        },
+        background_color,
         ..default()
     };
-    let hidden_frame_base = NodeBundle { style, background_color, ..default() };
 
-    let style = Style
-    {   width,
-        height,
-        position_type  : PositionType::Absolute,
-        flex_direction : FlexDirection::Column,
-        justify_content: JustifyContent::FlexStart, //画面の上端
+    let hidden_frame_middle = NodeBundle
+    {   style: Style
+        {   width,
+            height,
+            position_type  : PositionType::Absolute,
+            flex_direction : FlexDirection::Column,
+            justify_content: JustifyContent::Center, //画面の中央
+            align_items    : AlignItems::Center,     //中央揃え
+            ..default()
+        },
+        background_color,
         ..default()
     };
-    let hidden_frame_header = NodeBundle { style, background_color, ..default() };
 
-//     let hidden_frame_middle = NodeBundle
-//     {   style: Style
-//         {   flex_direction: FlexDirection::Column,
-//             align_items   : AlignItems::Center,
-//             ..default()
-//         },
-//         background_color,
-//         ..default()
-//     };
-
-    let style = Style
-    {   width,
-        height,
-        position_type  : PositionType::Absolute,
-        flex_direction : FlexDirection::Column,
-        justify_content: JustifyContent::FlexEnd, //画面の下端
+    let hidden_frame_footer = NodeBundle
+    {   style: Style
+        {   width,
+            height,
+            position_type  : PositionType::Absolute,
+            flex_direction : FlexDirection::Column,
+            justify_content: JustifyContent::FlexEnd, //画面の下端
+            ..default()
+        },
+        background_color,
         ..default()
     };
-    let hidden_frame_footer = NodeBundle { style, background_color, ..default() };
 
     //ヘッダーの準備
-    let mut header_left   = text_ui( &TEXT_HEADER_LEFT  , &asset_svr );
-    let mut header_center = text_ui( &TEXT_HEADER_CENTER, &asset_svr );
-    let mut header_right  = text_ui( &TEXT_HEADER_RIGHT , &asset_svr );
+    let mut header_left   = misc::text_ui( &TEXT_HEADER_LEFT  , &asset_svr );
+    let mut header_center = misc::text_ui( &TEXT_HEADER_CENTER, &asset_svr );
+    let mut header_right  = misc::text_ui( &TEXT_HEADER_RIGHT , &asset_svr );
     header_left.style.align_self   = AlignSelf::FlexStart;
     header_center.style.align_self = AlignSelf::Center;
     header_right.style.align_self  = AlignSelf::FlexEnd;
 
-
     //フッターの準備
-    let mut footer_left   = text_ui( &TEXT_FOOTER_LEFT  , &asset_svr );
-    let mut footer_center = text_ui( &TEXT_FOOTER_CENTER, &asset_svr );
-    let mut footer_right  = text_ui( &TEXT_FOOTER_RIGHT , &asset_svr );
+    let mut footer_left   = misc::text_ui( &TEXT_FOOTER_LEFT  , &asset_svr );
+    let mut footer_center = misc::text_ui( &TEXT_FOOTER_CENTER, &asset_svr );
+    let mut footer_right  = misc::text_ui( &TEXT_FOOTER_RIGHT , &asset_svr );
     footer_left.style.align_self   = AlignSelf::FlexStart;
     footer_center.style.align_self = AlignSelf::Center;
     footer_right.style.align_self  = AlignSelf::FlexEnd;
 
-    //隠しフレームの中に子要素を作成する
-    cmds.spawn( hidden_frame_base ).with_children
+    //隠しフレームと子要素を作成する
+    cmds.spawn( hidden_frame_header ).with_children
     (   | cmds |
-        {   cmds.spawn( hidden_frame_header ).with_children
-            (   | cmds |
-                {   //ヘッダー
-                    cmds.spawn( ( header_left  , HeaderUiStages  ) );
-                    cmds.spawn( ( header_center, HeaderUiScore   ) );
-                    cmds.spawn( ( header_right , HeaderUiHiScore ) );
-                }
-            );
-            cmds.spawn( hidden_frame_footer ).with_children
-            (   | cmds |
-                {   //フッター
-                    cmds.spawn( ( footer_left, FooterUiFps ) );
-                    cmds.spawn(   footer_center              );
-                    cmds.spawn(   footer_right               );
-                }
-            );
+        {   cmds.spawn( ( header_left  , UiStage   ) );
+            cmds.spawn( ( header_center, UiScore   ) );
+            cmds.spawn( ( header_right , UiHiScore ) );
         }
     );
+    cmds.spawn( ( hidden_frame_middle, HiddenFrameMiddle ) );
+    cmds.spawn( hidden_frame_footer ).with_children
+    (   | cmds |
+        {   cmds.spawn( ( footer_left, UiFps ) );
+            cmds.spawn(   footer_center        );
+            cmds.spawn(   footer_right         );
+        }
+    );
+
 
     //おまけ(蟹)
     let custom_size = Some ( SIZE_GRID * MAGNIFY_SPRITE_KANI );
@@ -257,34 +234,11 @@ fn spawn_text_ui
     ;
 }
 
-//TextBundleを作る
-fn text_ui
-(   message: &[ MessageSect ],
-    asset_svr: &Res<AssetServer>,
-) -> TextBundle
-{   let mut sections = Vec::new();
-    for ( line, file, size, color ) in message.iter()
-    {   let value = line.to_string();
-        let style = TextStyle
-        {   font     : asset_svr.load( *file ),
-            font_size: *size,
-            color    : *color
-        };
-        sections.push( TextSection { value, style } );
-    }
-    let alignment = TextAlignment::Center;
-    let position_type = PositionType::Absolute;
-
-    let text  = Text { sections, alignment, ..default() };
-    let style = Style { position_type, ..default() };
-    TextBundle { text, style, ..default() }
-}
-
 ////////////////////////////////////////////////////////////////////////////////
 
 //フッターを更新する(FPS)
 fn update_fps
-(   mut qry_text: Query<&mut Text, With<FooterUiFps>>,
+(   mut qry_text: Query<&mut Text, With<UiFps>>,
     diag_store: Res<DiagnosticsStore>,
 )
 {   let Ok( mut text ) = qry_text.get_single_mut() else { return };
@@ -308,45 +262,6 @@ fn update_fps
 ////////////////////////////////////////////////////////////////////////////////
 
 //End of code.
-
-
-
-
-
-// ////////////////////////////////////////////////////////////////////////////////
-
-// //プラグインの設定
-// pub struct Schedule;
-// impl Plugin for Schedule
-// {   fn build( &self, app: &mut App )
-//     {   app
-//         //ResourceとEvent
-//         .add_state::<MyState>()         //State登録
-//         .init_resource::<Record>()      //スコア等の初期化
-//         .init_resource::<Map>()         //迷路情報の初期化
-//         .init_resource::<CountDown>()   //カウントダウンタイマーの初期化
-//         .add_event::<EventClear>()      //ステージクリアイベントの登録
-//         .add_event::<EventOver>()       //ゲームオーバーイベントの登録
-//         .init_resource::<NowGamepad>()  //操作を受け付けるgamepad
-//         .init_resource::<CrossButton>() //十字ボタンの入力状態保存用
-
-//         //ゲームの準備
-//         .add_plugins( load_assets::Schedule ) //assetsの事前読込
-
-//         //Assetsの読込後
-//         .add_systems
-//         (   OnExit ( MyState::InitApp ),
-//             (   spawn_screen_frame, //ゲーム枠をspawn
-//                 spawn_text_ui,      //Text UIをspawn
-//             )
-//         )
-
-//         //FPSの表示
-//         .add_plugins( FrameTimeDiagnosticsPlugin ) //FPSプラグイン
-//         .add_systems( Update, update_footer_fps ) //表示更新
-//         ;
-//     }
-// }
 
 
 // ////////////////////////////////////////////////////////////////////////////////
@@ -480,17 +395,6 @@ fn update_fps
 //             );
 //         }
 //     );
-
-//     //おまけ(蟹)
-//     let px2d = Grid::new( SCREEN_GRIDS_WIDTH - 4, SCREEN_GRIDS_HEIGHT - 1 ).px2d_screen();
-//     let px3d = px2d.extend( DEPTH_SPRITE_KANI_DOTOWN );
-//     let custom_size = Some ( SIZE_GRID * MAGNIFY_SPRITE_KANI );
-//     cmds
-//     .spawn( SpriteBundle::default() )
-//     .insert( asset_svr.load( ASSETS_SPRITE_KANI_DOTOWN ) as Handle<Image> )
-//     .insert( Sprite { custom_size, ..default() } )
-//     .insert( Transform::from_translation( px3d ) )
-//     ;
 // }
 
 // ////////////////////////////////////////////////////////////////////////////////
