+++ src/play_game/mod.rs
@@ -1,0 +1,30 @@
+use super::*;
+
+//ゲームロジック
+mod schedule;
+pub use schedule::Schedule;
+
+//ヘッダーの情報表示
+mod header;
+
+//pause処理
+mod pause;
+
+//マップの処理
+pub mod map; //title_demoから呼び出すためpub付
+
+// //アプリの設定
+// mod config;
+// pub use config::*;
+
+// //型定義
+// mod types;
+// pub use types::*;
+
+// //ユーティリティ
+// pub mod misc;
+
+// //debug用
+// pub mod debug;
+
+//End of code.
