+++ src/public/config.rs
@@ -16,7 +16,7 @@ pub static MAIN_WINDOW: Lazy<Option<Window>> = Lazy::new
         {   title,
             resolution: ( SCREEN_PIXELS_WIDTH, SCREEN_PIXELS_HEIGHT ).into(),
             resizable: false,
-            // fit_canvas_to_parent: true, //不具合が発生した場合コメントアウトする
+            fit_canvas_to_parent: true, //不具合が発生した場合コメントアウトする
             ..default()
         };
 
@@ -101,23 +101,26 @@ counted_array!
 pub const DEPTH_SPRITE_DEBUG_GRID : f32 = 999.0; //重なりの最大値
 pub const DEPTH_SPRITE_KANI_DOTOWN: f32 = 900.0;
 pub const DEPTH_SPRITE_GAME_FRAME : f32 = 800.0;
-pub const DEPTH_SPRITE_CHASER     : f32 = 700.0;
-pub const DEPTH_SPRITE_PLAYER     : f32 = 600.0;
+// pub const DEPTH_SPRITE_CHASER     : f32 = 700.0;
+// pub const DEPTH_SPRITE_PLAYER     : f32 = 600.0;
 pub const DEPTH_SPRITE_DOT        : f32 = 500.0;
 pub const DEPTH_SPRITE_BRICK_WALL : f32 = 400.0;
 
-// //TEXT UI の重なり順
+//TEXT UI の重なり順
 // pub const ZINDEX_TEXTUI_PAUSE: ZIndex = ZIndex::Global ( 999 );
 
-// //スプライト色
-// pub const COLOR_SPRITE_DOT   : Color = Color::rgb( 1.0, 1.0, 0.7 );
+//スプライト色
+pub const COLOR_SPRITE_DOT   : Color = Color::rgb( 1.0, 1.0, 0.7 );
 // pub const COLOR_SPRITE_PLAYER: Color = Color::YELLOW;
 
-// //スプライト拡縮率
-// pub const MAGNIFY_SPRITE_DOT   : f32 = 0.08;
+//スプライト拡縮
+pub const MAGNIFY_SPRITE_DOT   : f32 = 0.08;
 // pub const MAGNIFY_SPRITE_PLAYER: f32 = 0.4;
 // pub const MAGNIFY_SPRITE_CHASER: f32 = 0.5;
 // pub const MAGNIFY_SPRITE_KANI  : f32 = 0.9;
+
+//調整値
+pub const ADJUSTER_MAP_SPRITES: Vec2 = Vec2::new( 0.0, 1.0 * PIXELS_PER_GRID );
 
 // //SEボリューム
 // pub const VOLUME_SOUND_BEEP: f32 = 0.1;
@@ -172,6 +175,23 @@ const ER_BAD_SCREEN_DESIGN: &str = "Frame design unmatch width/height parameters
 
 ////////////////////////////////////////////////////////////////////////////////
 
+//PAUSEのキー／ボタン
+pub const PAUSE_KEY: KeyCode = KeyCode::Escape;
+pub const PAUSE_BUTTON: GamepadButtonType = GamepadButtonType::Select; //PS4[SHARE]
+
+////////////////////////////////////////////////////////////////////////////////
+
+//フルスクリーンのキー／ボタン
+pub const FULL_SCREEN_KEY: KeyCode = KeyCode::Return;
+counted_array!
+(   pub const FULL_SCREEN_KEY_MODIFIER: [ KeyCode; _ ] =
+    [   KeyCode::AltRight,
+        KeyCode::AltLeft,
+    ]
+);
+pub const FULL_SCREEN_BUTTON: GamepadButtonType = GamepadButtonType::Start; //ps4[OPTION]
+
+////////////////////////////////////////////////////////////////////////////////
 //マップ縦横幅(Grid)
 pub const MAP_GRIDS_WIDTH : i32 = SCREEN_GRIDS_WIDTH;
 pub const MAP_GRIDS_HEIGHT: i32 = SCREEN_GRIDS_HEIGHT - 2;
@@ -181,19 +201,26 @@ pub const MAP_GRIDS_X_RANGE: Range<i32> = 0..MAP_GRIDS_WIDTH;
 pub const MAP_GRIDS_Y_RANGE: Range<i32> = 0..MAP_GRIDS_HEIGHT;
 
 //外壁を含まないレンジ
-pub const MAP_GRIDS_X_RANGE_INNER: Range<i32> = 1..MAP_GRIDS_WIDTH  - 1;
-pub const MAP_GRIDS_Y_RANGE_INNER: Range<i32> = 1..MAP_GRIDS_HEIGHT - 1;
+// pub const MAP_GRIDS_X_RANGE_INNER: Range<i32> = 1..MAP_GRIDS_WIDTH  - 1;
+// pub const MAP_GRIDS_Y_RANGE_INNER: Range<i32> = 1..MAP_GRIDS_HEIGHT - 1;
 
 ////////////////////////////////////////////////////////////////////////////////
 
+//End of code.
+
+
+
+////////////////////////////////////////////////////////////////////////////////
+
+
 //四方の配列
-pub const NEWS: [ News; 4 ] = [ News::South, News::East, News::West, News::North ];
+// pub const NEWS: [ News; 4 ] = [ News::South, News::East, News::West, News::North ];
 
 ////////////////////////////////////////////////////////////////////////////////
 
 //Playerの設定値
-pub const PLAYER_TURN_COEF: f32 = 3.5;
-pub const PLAYER_MOVE_COEF: f32 = 3.5;
+// pub const PLAYER_TURN_COEF: f32 = 3.5;
+// pub const PLAYER_MOVE_COEF: f32 = 3.5;
 
 // pub const UNIT_TURN: f32 = FRAC_PI_2;
 // pub const UNIT_MOVE: f32 = 1.0;
@@ -202,35 +229,23 @@ pub const PLAYER_MOVE_COEF: f32 = 3.5;
 // pub const PLAYER_MOVE_COEF: f32 = PIXELS_PER_GRID / PLAYER_WAIT; //移動の中割係数
 
 //Chaserの設定値
-pub const CHASER_WAIT: f32 = 0.13;                               //移動のウェイト
-pub const CHASER_MOVE_COEF: f32 = PIXELS_PER_GRID / CHASER_WAIT; //移動の中割係数
-pub const CHASER_ACCEL: f32 = 0.4;                               //スピードアップの割増
-pub const CHASER_INIT_POSITION: [ ( i32, i32 ); 4 ] =            //スタート座標(Grid)
-[   ( 1    , 1     ),
-    ( 1    , MAX_Y ),
-    ( MAX_X, 1     ),
-    ( MAX_X, MAX_Y ),
-];
-const MAX_X: i32 = MAP_GRIDS_WIDTH  - 2;
-const MAX_Y: i32 = MAP_GRIDS_HEIGHT - 2;
+// pub const CHASER_WAIT: f32 = 0.13;                               //移動のウェイト
+// pub const CHASER_MOVE_COEF: f32 = PIXELS_PER_GRID / CHASER_WAIT; //移動の中割係数
+// pub const CHASER_ACCEL: f32 = 0.4;                               //スピードアップの割増
+// pub const CHASER_INIT_POSITION: [ ( i32, i32 ); 4 ] =            //スタート座標(Grid)
+// [   ( 1    , 1     ),
+//     ( 1    , MAX_Y ),
+//     ( MAX_X, 1     ),
+//     ( MAX_X, MAX_Y ),
+// ];
+// const MAX_X: i32 = MAP_GRIDS_WIDTH  - 2;
+// const MAX_Y: i32 = MAP_GRIDS_HEIGHT - 2;
 
 ////////////////////////////////////////////////////////////////////////////////
 
 
 // ////////////////////////////////////////////////////////////////////////////////
 
-// //フルスクリーン切替のキーとパッドボタン
-// counted_array!
-// (   pub const FULL_SCREEN_KEYS: [ ( Option<KeyCode>, KeyCode ); _ ] =
-//     [   ( Some ( KeyCode::AltLeft  ), KeyCode::Return ),
-//         ( Some ( KeyCode::AltRight ), KeyCode::Return ),
-//     ]
-// );
-// pub const FULL_SCREEN_BUTTON: GamepadButtonType = GamepadButtonType::Select; //PS4:SHARE
-
-// //PAUSEのキーとパッドボタン
-// pub const PAUSE_KEY: KeyCode = KeyCode::Escape;
-// pub const PAUSE_BUTTON: GamepadButtonType = GamepadButtonType::Start; //PS4:OPTIONS
 
 // //Hit ANY Keyの処理で押されても無視するキーとボタン
 // counted_array!
@@ -321,55 +336,6 @@ const MAX_Y: i32 = MAP_GRIDS_HEIGHT - 2;
 // );
 // pub const TEXT_UI_OVER: TextUiOver = TextUiOver( 10, MyState::Title, 6, cd_string_over, MyState::StageStart );
 // fn cd_string_over( n: i32 ) -> String { n.to_string() }
-
-// //text UIの設定値
-// #[allow( dead_code )]
-// pub const NA2  : &str = "##";
-// pub const NA3  : &str = "###";
-// pub const NA5  : &str = "#####";
-// pub const NA2_5: &str = "##-#####";
-// pub const NA3_2: &str = "###.##";
-
-// //ヘッダーに表示するtext UI
-// counted_array!
-// (   pub const HEADER_LEFT_TEXT: [ MessageSect; _ ] =
-//     [   ( " STAGE ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.7, Color::GOLD  ),
-//         ( NA2      , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.7, Color::WHITE ),
-//     ]
-// );
-// counted_array!
-// (   pub const HEADER_CENTER_TEXT: [ MessageSect; _ ] =
-//     [   ( " SCORE ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.7, Color::GOLD   ),
-//         ( NA5      , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.7, Color::WHITE  ),
-//         ( NA3      , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.5, Color::SILVER ),  //placeholder for debug
-//     ]
-// );
-// counted_array!
-// (   pub const HEADER_RIGHT_TEXT: [ MessageSect; _ ] =
-//     [   ( " Hi-SCORE ", ASSETS_FONT_ORBITRON_BLACK      , PIXELS_PER_GRID * 0.7, Color::GOLD  ),
-//         ( NA5         , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.7, Color::WHITE ),
-//     ]
-// );
-
-// //フッターに表示するtext UI
-// counted_array!
-// (   pub const FOOTER_LEFT_TEXT: [ MessageSect; _ ] =
-//     [   ( " FPS " , ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID       * 0.6, Color::TEAL   ),
-//         ( NA3_2   , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.4, Color::SILVER ),
-//         ( " demo ", ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID       * 0.6, Color::TEAL   ),
-//         ( NA2_5   , ASSETS_FONT_PRESSSTART2P_REGULAR, PIXELS_PER_GRID * 0.4, Color::SILVER ),
-//     ]
-// );
-// counted_array!
-// (   pub const FOOTER_CENTER_TEXT: [ MessageSect; _ ] =
-//     [   ( "hyoi 2021 - 2023", ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID * 0.6, Color::TEAL ),
-//     ]
-// );
-// counted_array!
-// (   pub const FOOTER_RIGHT_TEXT: [ MessageSect; _ ] =
-//     [   ( "Powered by RUST & BEVY ", ASSETS_FONT_ORBITRON_BLACK, PIXELS_PER_GRID * 0.6, Color::TEAL ),
-//     ]
-// );
 
 // ////////////////////////////////////////////////////////////////////////////////
